46_bilboard_spotify_playlist/requirements.txt
bs4==0.0.2
requests==2.32.3
python-dotenv==1.0.1
spotipy==2.24.0